requirements.txt
chatterbot
chatterbot_corpus
slackclient
websocket-client
beautifulsoup4